+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Fixed a bug when cross-referencing components caused infinite loop in Pull API (isuue #58, pull request #61)
 - Pull API handles recursion more like Datomic does (issue #62)
 - Exposed filtered API to js (pull request #65)
+- Added ICounted, ISequable, IEmptyableCollection to `datascript.core/DB`
+- Force put `Datom` and `DB` tag readers to `cljs.reader/*tag-table*`, do not rely on `data_readers.clj`
 
 # 0.9.0